+++ src/Auth/introspection.js
@@ -1,5 +1,5 @@
 const fetch = require("node-fetch");
-const config = require("./config");
+const config = require("../config");
 
 async function verifyToken(request){
